[FEAT] 채팅 관련 엔티티 생성
