+++ src/main/java/com/beautiflow/chat/domain/TemplateMessage.java
@@ -1,10 +1,7 @@
-package com.beautiflow.chat;
+package com.beautiflow.chat.domain;
 
-import java.time.LocalDateTime;
-
-import com.beautiflow.global.domain.BaseTimeEntity;
-import com.beautiflow.global.domain.SenderType;
-import com.beautiflow.user.domain.User;
+import com.beautiflow.global.domain.SendTimeType;
+import com.beautiflow.shop.domain.Shop;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.EnumType;
@@ -26,21 +23,21 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @Builder
 @Entity
-@Table(name = "messages")
-public class ChatMessage extends BaseTimeEntity {
+@Table(name = "template_messages")
+public class TemplateMessage {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	private Long id;
 
 	@ManyToOne(fetch = FetchType.LAZY)
-	private ChatRoom chatRoom;
+	private Shop shop;
 
-	@ManyToOne(fetch = FetchType.LAZY)
-	private User sender;
+	private String title;
+	private String content;
+	private boolean enabled;
 
 	@Enumerated(EnumType.STRING)
-	private SenderType senderType;
+	private SendTimeType sendTimeType;
 
-	private String content;
-	private String imageUrl;
+	private String tag;
 }
